finish rewrite & xcm test
