--- pallets/xtransfer/src/xcm/mod.rs
@@ -1,5 +1,0 @@
-pub mod xcm_helper;
-pub mod xcm_transfer;
-
-#[cfg(test)]
-mod mock;
